Refactor: CTS: merge crm_common class into ClusterManager class
